requirements.txt
pytesseract
pillow
pyperclip
keyboard